Rename `BlendMode` -> `BlendModes`
